Split the mod picker seasonal section by season. Messed around with some styling.
