--- src/app/loadout-builder/filter/PerksForBucket.m.scss
@@ -1,14 +1,0 @@
-.perks {
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(250px, 1fr));
-  gap: 10px;
-  margin-bottom: 16px;
-}
-
-.bucket {
-  h3 {
-    text-transform: uppercase;
-    letter-spacing: 1px;
-    margin: 0 0 4px 0;
-  }
-}
